requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
babel==2.18.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.1.8
courlan==1.3.2
dateparser==1.2.2
exceptiongroup==1.3.1
fastapi==0.128.8
h11==0.16.0
htmldate==1.9.4
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jusText==3.0.2
lxml==6.0.2
lxml_html_clean==0.4.4
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
regex==2026.1.15
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.49.3
tavily-python==0.7.23
tiktoken==0.12.0
tld==0.13.2
trafilatura==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.39.0
